refactor(consumption): extract Fuel/Charging tabs from consumption_screen (Refs #563 phase: consumption_screen)
Pure extraction of the two oversized inline tab bodies on the
ConsumptionScreen so the parent screen drops below the 300-LOC ceiling
and each child file owns one concern.

- consumption_screen.dart 481 -> 269 LOC (parent, navigation + FAB)
- fuel_tab.dart 87 LOC (new — fill-up list, stats, help banner)
- charging_tab.dart 145 LOC (new — charging log list + charts header)

Behaviour unchanged: same widget tree, same Dismissible keys, same
Provider reads, same FAB rebinding logic. The previously private
_FuelTab / _ChargingTab classes had no external test references, so
promoting them to public FuelTab / ChargingTab doesn't break anything.

Existing 19 consumption_screen_*_test.dart tests pass; related
widget tests (charging_charts, charging_log_card, fill_up_card,
consumption_stats_card) also green.
